+++ src/Data/Morpheus/Execution/Document/Encode.hs
@@ -20,7 +20,7 @@ import           Data.Morpheus.Execution.Server.Encode
                                                 , ObjectResolvers(..)
                                                 )
 import           Data.Morpheus.Types.GQLType    ( TRUE )
-import           Data.Morpheus.Types.Internal.AST.Data
+import           Data.Morpheus.Types.Internal.AST
                                                 ( DataField(..)
                                                 , QUERY
                                                 , SUBSCRIPTION