+++ src/Data/Morpheus/Execution/Internal/Decode.hs
@@ -26,16 +26,15 @@ import           Language.Haskell.TH            ( ExpQ
 import           Data.Morpheus.Error.Internal   ( internalArgumentError
                                                 , internalTypeMismatch
                                                 )
-import           Data.Morpheus.Types.Internal.AST.Data
+import           Data.Morpheus.Types.Internal.AST
                                                 ( DataField(..)
                                                 , Key
-                                                , ConsD(..) )
+                                                , ConsD(..) 
+                                                , Object
+                                                , Value(..))
 import           Data.Morpheus.Types.Internal.Resolving
                                                 ( Validation )
-import           Data.Morpheus.Types.Internal.AST.Value
-                                                ( Object
-                                                , Value(..)
-                                                )
+
 
 decodeObjectExpQ :: ExpQ -> ConsD -> ExpQ
 decodeObjectExpQ fieldDecoder ConsD { cName, cFields } = handleFields cFields