+++ src/Data/Morpheus/Parsing/Request/Parser.hs
@@ -34,24 +34,25 @@ import           Data.Morpheus.Types.Internal.Resolving
                                                 ( Validation
                                                 , Failure(..)
                                                 )
-import           Data.Morpheus.Types.Internal.AST.Value
+import           Data.Morpheus.Types.Internal.AST
                                                 ( Value(..)
                                                 , replaceValue
+                                                , GQLQuery(..)
                                                 )
 import           Data.Morpheus.Types.IO         ( GQLRequest(..) )
-import           Data.Morpheus.Types.Types      ( GQLQueryRoot(..) )
 
-parseGQLSyntax :: Text -> Either (ParseErrorBundle Text Void) GQLQueryRoot
+
+parseGQLSyntax :: Text -> Either (ParseErrorBundle Text Void) GQLQuery
 parseGQLSyntax = runParser request "<input>"
  where
-  request :: Parser GQLQueryRoot
-  request = label "GQLQueryRoot" $ do
+  request :: Parser GQLQuery
+  request = label "GQLQuery" $ do
     spaceAndComments
     operation <- parseOperation
     fragments <- manyTill parseFragmentDefinition eof
-    pure GQLQueryRoot { operation, fragments, inputVariables = [] }
+    pure GQLQuery { operation, fragments, inputVariables = [] }
 
-parseGQL :: GQLRequest -> Validation GQLQueryRoot
+parseGQL :: GQLRequest -> Validation GQLQuery
 parseGQL GQLRequest { query, variables } = case parseGQLSyntax query of
   Right root       -> pure $ root { inputVariables = toVariableMap variables }
   Left  parseError -> failure $ processErrorBundle parseError