+++ src/Data/Morpheus/Types.hs
@@ -39,10 +39,11 @@ import           Data.Morpheus.Types.GQLScalar  ( GQLScalar
                                                 )
 import           Data.Morpheus.Types.GQLType    ( GQLType(KIND, description) )
 import           Data.Morpheus.Types.ID         ( ID(..) )
-import           Data.Morpheus.Types.Internal.AST.Data
+import           Data.Morpheus.Types.Internal.AST
                                                 ( MUTATION
                                                 , QUERY
                                                 , SUBSCRIPTION
+                                                , ScalarValue(..)
                                                 )
 import           Data.Morpheus.Types.Internal.Resolving
                                                 ( Event(..)
@@ -51,8 +52,6 @@ import           Data.Morpheus.Types.Internal.Resolving
                                                 , LiftEither(..)
                                                 , lift
                                                 )
-import           Data.Morpheus.Types.Internal.AST.Value
-                                                ( ScalarValue(..) )
 import           Data.Morpheus.Types.IO         ( GQLRequest(..)
                                                 , GQLResponse(..)
                                                 )